+++ ehcore/src/main/java/com/everhomes/server/schema/tables/daos/EhBizCouponDao.java
@@ -85,7 +85,7 @@ public class EhBizCouponDao extends org.jooq.impl.DAOImpl<com.everhomes.server.s
 	/**
 	 * Fetch records that have <code>status IN (values)</code>
 	 */
-	public java.util.List<com.everhomes.server.schema.tables.pojos.EhBizCoupon> fetchByStatus(java.lang.Integer... values) {
+	public java.util.List<com.everhomes.server.schema.tables.pojos.EhBizCoupon> fetchByStatus(java.lang.Byte... values) {
 		return fetch(com.everhomes.server.schema.tables.EhBizCoupon.EH_BIZ_COUPON.STATUS, values);
 	}
